Add files via upload
One line Description of Project
Design and Develop a Console based Hotel Management System.

Overview of Project
This project will facilitate the Hotel owner in the Hotel to keep record of rooms and reservations and create a report on regular basis.

Project Detail
The Hotel for which you are creating this project has 5 floors and each floor has 50 rooms.
Following are types of rooms:

1.	Standard
a.	This category usually means the most basic room type offered by the hotel. It has basic, standard amenities and furnishings. A standard room in a Four Seasons hotel is without question much more deluxe than a standard in, say, a Holiday Inn, but there may be higher categories from which to choose. Standard rooms in hotels with higher categories often have no view or have a poor view over the dumpster or parking lot.
b.	Price is Rs.300 / 24 Hours
c.	This Hotel contains total of 50 rooms of this type, 10 on each floor.
2.	Moderate
a.	Usually a slight bit better than standard, but still not deluxe. It may refer to the room view as well as the size and type of furnishings offered.
b.	Price is Rs. 500 / 24 Hours
c.	This Hotel contains total of 50 rooms of this type, 10 on each floor.
3.	Superior
a.	This category is always subject to interpretation. It's supposed to mean superior to a standard room in both size and furnishings, but it often refers to just the view. Some hotels have only Superior rooms; the categories then are defined by the view and location of the room.
b.	Price is Rs. 1000 / 24 Hours
c.	This Hotel contains total of 50 rooms of this type, 10 on each floor.
4.	Junior Suite
 
a.	A "junior" suite is typically a larger room with a separate seating area. Sometimes it's got a small divider between the part of the room that the bed is in and the seating area, but it is not two separate rooms.
b.	Price is Rs. 2000 / 24 Hours
c.	This Hotel contains total of 50 rooms of this type, 10 on each floor.
5.	Suite
a.	A Suite is usually two or more rooms clearly defined; a bedroom and a living or sitting room, with a door that closes between them. Many hotels use the word "suite" to define any room with a sofa in it so be sure to check thoroughly if what you really want are the two or more separate rooms.
b.	Price is Rs. 5000 / 24 Hour
c.	This Hotel contains total of 50 rooms of this type, 10 on each floor.

This Hotel records following information of its customer

•	Full Name

•	Age

•	Gender

•	ID card-Number

•	Balance in Rupees (If paid in advance, balance will be zero)

•	Total Days to reserve

•	Floor Number on which reserved room is

•	Room Type

•	Room Number

•	Check in time and date

•	Checkout time and date

•	Time remaining in the Hotel ( in Hours, should be calculated on runtime, automatically)

Menu/Option










•	If room is reserved it should not be reserved again until customer checkout or reserve time is ended.

•	When room is reserved, it should ask customer if he/she is old customer or new customer.

o If new customer, then record all the above mentioned things and give customer a room.

o If old customer, then search old customer name in the database (file) and retrieve all his/her info and give him/her a room.

•	NOTE: There is a lot of difference in Reserve and Check in a customer.

•	In Detail Report you’ve to give daily base statistics for example
o How many customers have checked in today o How many have checked out today

o How many rooms have been reserved and their types o How many are empty and their types.
 
•	You also have to save all the information in the files. There is no restriction on the file format.

Class Design

For Rooms make following class hierarchy:

Room









Standard		Moderate		Superior		Junior Suite		Suite
								

You must use all concept of classes including inheritance, polymorphism along with arrays and functions etc.


Data Storage

1)	In Files

•	All the data should be saved in files.

•	Filing section contains almost 50 % marks of the whole project.

•	If I close the program and run it again, it should read all the records and old data that were saved last time.

•	You can use any format for filing.

•	All the information of room should be saved in rooms.dat file

•	All the information of customers should be saved in customers.dat file